KynZo
-Add Req knplabs/knp-time-bundle
-Add Req twig/string-extra
-Add new migration ( createdAt + UpdatedAt )
-Add favicon
-Edit Home controller ( add ProjetsRepository )
-Edit Security.yaml ( add remember_me + access_control + role_hierarchy)
-Edit Twig.yaml ( add form_themes )
-Edit style.css ( Remove Flotter css )
-Edit LoginController.php ( Active eedirectToRoute )
-Edit ProjetsController.php ( add filter for projetsRepository + grantedAccess + remove Flash)
-Edit Projets.php ( Add CreatedAt + Add UpdatedAt + Add Timestamps ( HasLifecycleCallbacks() ) )
-Edit RegistrationFormType ( Add LastName + Add FirstName + Birthdays )
-Edit LoginAuthenticator (Add redirectResponse )
-Edit base.html.twig ( lang fr )

+++ src/Controller/ProjetsController.php
@@ -17,13 +17,15 @@ class ProjetsController extends AbstractController
     public function index(ProjetsRepository $projetsRepository): Response
     {
         return $this->render('projets/index.html.twig', [
-            'projets' => $projetsRepository->findAll(),
+            'projets' => $projetsRepository->findBy([], ['CreatedAt' => 'DESC']),
         ]);
     }
 
     #[Route('/new', name: 'projets_new', methods: ['GET','POST'])]
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $projet = new Projets();
         $form = $this->createForm(ProjetsType::class, $projet);
         $form->handleRequest($request);
@@ -32,8 +34,6 @@ class ProjetsController extends AbstractController
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($projet);
             $entityManager->flush();
-
-            $this->addFlash('success', 'Projet créé avec succès');
 
             return $this->redirectToRoute('projets_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -55,13 +55,13 @@ class ProjetsController extends AbstractController
     #[Route('/{id}/edit', name: 'projets_edit', methods: ['GET','POST'])]
     public function edit(Request $request, Projets $projet): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $form = $this->createForm(ProjetsType::class, $projet);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
-            
-            $this->addFash('success', 'Projet modifié avec succès');
 
             return $this->redirectToRoute('projets_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -75,12 +75,13 @@ class ProjetsController extends AbstractController
     #[Route('/{id}', name: 'projets_delete', methods: ['POST'])]
     public function delete(Request $request, Projets $projet): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+        
         if ($this->isCsrfTokenValid('delete'.$projet->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($projet);
             $entityManager->flush();
 
-            $this->addFash('notice', 'Projet supprimé avec succès');
         }
 
         return $this->redirectToRoute('projets_index', [], Response::HTTP_SEE_OTHER);
